Add App
Add basic app utilizing Xamarin an Visual Studio. This should allow the app to be used with both IOS and Android. Thus far the app has been successfully emulated on Android. The app consists of a home page with a few buttons to navigate around to other blank pages. Next up is RFID and Bluetooth connectivity.